changed initialization and added calling of the Registry to the ApplicationConfig class. Finally ....
